refactor: colocate virtual flatten rows hook
